requirements.txt
uvicorn==0.29.0
fastapi==0.105.0
fastapi-apscheduler==0.0.5
httpx==0.27.0
beautifulsoup4==4.12.2
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
mysql-connector-python==8.4.0
python-multipart==0.0.9